+++ src/indexing.jl
@@ -1,0 +1,3 @@
+import Base: getindex
+getindex(od::OddFrame, col::Symbol) = od.lookup[col]
+getindex(od::OddFrame, col::String) = od.lookup[Symbol(col)]
